requirements.txt
flask
python-dotenv
PyPDF2
fpdf2
requests
supabase
google-generativeai
gunicorn